fix: [TEMP] Remove Rank
This is a temporary workaround until final decision was made regarding my profile.

There is an improvement in algorithm of Rank calculation for https://github.com/anuraghazra/github-readme-stats. While calculation is now done differently, this might be a issue if you are working on personal projects. It is possible that despite working on personal projects fully, that rank can be still be low.